% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmeta_get_translation_plot.R
\name{get_translation_data}
\alias{get_translation_data}
\title{Get Translation Data}
\usage{
get_translation_data(preliminary_analysis_data = NULL)
}
\arguments{
\item{preliminary_analysis_data}{A list containing multiple elements:
\itemize{
\item The 7th element (\code{txdb_features}) contains genomic feature data (e.g., exons).
\item The 2nd element (\code{peak_gr}) contains the genomic regions of interest (e.g., peaks).
\item The 9th element (\code{short_flank_size}) represents the flank size used for analysis.
\item The 10th element (\code{tx_length_file}) contains transcript length information for the genes.
}}
}
\value{
A list with two elements:
\enumerate{
\item A data table containing the coverage information for both translation start sites and end sites (\code{tsl_list}).
\item A list of peak heatmap data generated from the translation-related regions (\code{peak_heatmap_data}).
}
}
\description{
Get Translation Data
}
