% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmeta_get_translation_plot.R
\name{RNAmeta_get_translation_plot}
\alias{RNAmeta_get_translation_plot}
\title{Generate Translation Plot}
\usage{
RNAmeta_get_translation_plot(translation_data = NULL, set_group_name = NULL)
}
\arguments{
\item{translation_data}{A list containing:
\itemize{
\item \code{tsl_list}: A data table with translation start and end site coverage.
\item \code{peak_heatmap_data}: A list containing translation heatmap data.
}}

\item{set_group_name}{An optional vector of custom group names to replace the default sample names in the data.
If \code{NULL}, the function will use the original sample names.}
}
\value{
A list of four ggplot objects:
\enumerate{
\item Translation Start Site (TSS) Density Plot.
\item Translation End Site (TES) Density Plot.
\item Heatmap for the TSS region.
\item Heatmap for the TES region.
}
}
\description{
Generate Translation Plot
}
