% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmeta_peak_distribution_plot.R
\name{RNAmeta_peak_distribution_plot}
\alias{RNAmeta_peak_distribution_plot}
\title{Plot Peak Distribution}
\usage{
RNAmeta_peak_distribution_plot(
  preliminary_analysis_data = NULL,
  set_group_name = NULL,
  features_to_remove = NULL
)
}
\arguments{
\item{preliminary_analysis_data}{A list containing the preliminary analysis data, typically including genomic features and their frequencies across different samples.}

\item{set_group_name}{A vector of group names to be used for color grouping in the plot. The number of group names should match the number of unique samples in \code{all_sites}.}

\item{features_to_remove}{A vector of feature names (e.g., "Promoter", "UTR5", etc.) that should be excluded from the analysis and visualization.}
}
\value{
A \code{ggplot} object representing the bar plot of peak distribution across features. The plot shows the frequency of each feature for different samples, with a manual color scheme applied to each group.
}
\description{
Plot Peak Distribution
}
