% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmeta_peak_exon_type_plot.R
\name{RNAmeta_peak_exon_type_plot}
\alias{RNAmeta_peak_exon_type_plot}
\title{RNAmeta_peak_exon_type_plot}
\usage{
RNAmeta_peak_exon_type_plot(
  preliminary_analysis_data = NULL,
  set_group_name = NULL
)
}
\arguments{
\item{preliminary_analysis_data}{A list of data containing the peak exon type information. The data should contain exon information, with columns such as \code{Sample} and \code{exonLoc}.
Each element in the list corresponds to one set of data to be plotted.}

\item{set_group_name}{An optional vector of custom group names for the \code{Sample} column. If \code{NULL}, the function will use the original sample names in the data.
If provided, the length of this vector must match the number of unique sample groups.}
}
\value{
A \code{ggplot} object representing a bar plot of the percentage distribution of exon locations for each sample. The bars are grouped by \code{Sample} and colored by \code{exonLoc}.
The legend position will adjust based on the number of unique sample groups.
}
\description{
RNAmeta_peak_exon_type_plot
}
