% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmeta_peak_gene_size_plot.R
\name{RNAmeta_peak_gene_size_plot}
\alias{RNAmeta_peak_gene_size_plot}
\title{RNA Meta Peak Gene Size Plot}
\usage{
RNAmeta_peak_gene_size_plot(
  preliminary_analysis_data = NULL,
  set_group_name = NULL
)
}
\arguments{
\item{preliminary_analysis_data}{A list containing the preliminary analysis data. It should include the transcript length (\code{tx_len}) information extracted from the dataset.}

\item{set_group_name}{A vector of user-defined group names to rename the sample groups in the dataset.
The length of \code{set_group_name} must match the number of unique \code{Sample} values in the data.}
}
\value{
A \code{ggplot} object representing a box plot of transcript lengths (\code{tx_len}) across different sample groups.
\itemize{
\item The x-axis represents different sample groups.
\item The y-axis represents transcript length in log2 scale (\verb{log2(txLength bp)}).
\item The box plot shows the distribution of transcript lengths within each sample group.
\item The legend placement is determined based on the number of unique samples.
}
}
\description{
RNA Meta Peak Gene Size Plot
}
