% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmeta_meta_plot.R
\name{RNAmeta_species_list}
\alias{RNAmeta_species_list}
\title{Supported Species List}
\usage{
RNAmeta_species_list()
}
\value{
a data frame containing the full names of species.
}
\description{
This function generates a data frame containing the full names of several
model organisms and other species supported by the rnameta package.
}
